requirements.txt
# Core dependencies
requests>=2.31.0,<3.0.0
python-telegram-bot>=20.7,<21.0.0
python-dotenv>=1.0.0,<2.0.0

# Async HTTP client for concurrent requests
aiohttp>=3.9.0,<4.0.0

# Retry mechanism with exponential backoff
tenacity>=8.2.0,<9.0.0

# Type hints and data validation
typing-extensions>=4.8.0; python_version < "3.10"
